🚚 Improve class navigation
